scheduler/requirements.txt
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-redis==3.5.0
arrow==1.3.0
gitpython==3.1.40
more-itertools==10.1.0
nested-lookup==0.2.25
psycopg2-binary==2.9.9
